[robotInterface] Fix build with C++11
std::make_pair syntax is slightly changed in C++11
C++97 used to make a copy of the object, so it doesn't make any
difference if the passed objects are const.
C++11 uses rvalue references and move semantics, so the passed
objects are modified, and therefore the objects must be copied
explicitly.

The C++11 version works also for C++97, but it causes an extra
copy of the object, and for now I want to avoid this.

This fix is necessary because MSVC 2012 enables C++11 by
default.