EmbObjSkin: fix bug on eoSkin Callback
If robotInterface was started when the EMS boards are already sending data (for ex. after a robotInterface crash), 
the callback initialization was not properly done leaving the robotInterface unable to handle the skin messages.

Using the std::map, changing the use of the '[] 'operator with the '.at()' solved the issue.